chore: <- format

+++ src/InvestmentVault.sol
@@ -19,9 +19,9 @@ interface CErc20 {
 
     function supplyRatePerBlock() external returns (uint256);
 
-    function redeem(uint) external returns (uint);
+    function redeem(uint256) external returns (uint256);
 
-    function redeemUnderlying(uint) external returns (uint);
+    function redeemUnderlying(uint256) external returns (uint256);
 
     function balanceOf(address owner) external view returns (uint256 balance);
 }
